requirements.txt
python-dotenv
requests
google-generativeai
langdetect
ahocorasick
comet-ml
flask
gTTS